minor ups

+++ SampleArch.Model/Migrations/201603011359267_BNC.cs
@@ -44,7 +44,7 @@ namespace SampleArch.Model.Migrations
                 "dbo.Countries",
                 c => new
                     {
-                        Id = c.Int(nullable: false, identity: true),
+                        Id = c.Long(nullable: false, identity: true),
                         Name = c.String(nullable: false, maxLength: 100),
                     })
                 .PrimaryKey(t => t.Id);
@@ -122,7 +122,7 @@ namespace SampleArch.Model.Migrations
                         Phone = c.String(nullable: false, maxLength: 20),
                         Address = c.String(nullable: false, maxLength: 100),
                         State = c.String(nullable: false, maxLength: 50),
-                        CountryId = c.Int(nullable: false),
+                        CountryId = c.Long(nullable: false),
                         CreatedDate = c.DateTime(nullable: false),
                         CreatedBy = c.String(maxLength: 256),
                         UpdatedDate = c.DateTime(nullable: false),
